pyproject.toml
[project]
name = "sift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.66.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.2",
]
